nudges.asm
Routines to work out the best possible win with the nudges you have available.  I'm pretty pleased with this routine since it works out how to avoid triggering a lower value win when nudging the wheels.  The algorithm took a day or so to work out and optimise since there are millions of possible permutations on a maximum 10 nudge (up/down) on 5 wheels.  Processing time could take minutes, this algorithm takes a second.

(C)1990 Jason Brooks