Create Pacland.bas
Pacland had an interesting internal protection system once the game was loaded, clearly created on another machine (maybe PDS?) in that the final Xor was against the entire 64Kb in memory, which meant introducing additional code was impossible without using extra RAM, or working out a way to keep sync with the algorithm used.  None the less it was cracked! But little short of Multi-Face pokes, in memory I don't recall a cheat mode being published that used the original Cassette